Move DTOs into their own folder and merge CsvRequest with Request
