+++ src/Api/Endpoints/Dtos/ClearanceRequestsSummaryIntervalResponse.cs
@@ -1,5 +1,5 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
 public record ClearanceRequestsSummaryIntervalResponse([property: JsonPropertyName("unique")] int Unique);
