+++ src/Api/Endpoints/Dtos/DatumResponse.cs
@@ -1,5 +1,5 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
 public record DatumResponse<T>([property: JsonPropertyName("data")] IReadOnlyList<T> Data);
