+++ src/Api/Endpoints/Dtos/IntervalsResponse.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
 public record IntervalsResponse(
     [property: JsonPropertyName("releases")] IntervalsResponse<IntervalResponse<ReleasesSummaryResponse>> Releases,
