+++ src/Api/Endpoints/Dtos/LastReceivedMessageResponse.cs
@@ -1,8 +1,8 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
-public record ReleasesResponse(
+public record LastReceivedMessageResponse(
     [property: JsonPropertyName("timestamp")] DateTime Timestamp,
     [property: JsonPropertyName("reference")] string Reference
 );
