+++ src/Api/Endpoints/Dtos/MatchResponse.cs
@@ -1,8 +1,8 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
-public record LastReceivedMessageResponse(
+public record MatchResponse(
     [property: JsonPropertyName("timestamp")] DateTime Timestamp,
     [property: JsonPropertyName("reference")] string Reference
 );
