+++ src/Api/Endpoints/Dtos/MatchesSummaryResponse.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
 public record MatchesSummaryResponse(
     [property: JsonPropertyName("match")] int Match,
