+++ src/Api/Endpoints/Dtos/ReleasesResponse.cs
@@ -1,8 +1,8 @@
 using System.Text.Json.Serialization;
 
-namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+namespace Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 
-public record MatchResponse(
+public record ReleasesResponse(
     [property: JsonPropertyName("timestamp")] DateTime Timestamp,
     [property: JsonPropertyName("reference")] string Reference
 );
