+++ src/Api/Endpoints/ReleasesEndpoints.cs
@@ -1,4 +1,5 @@
 using Defra.TradeImportsReportingApi.Api.Data;
+using Defra.TradeImportsReportingApi.Api.Endpoints.Dtos;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Defra.TradeImportsReportingApi.Api.Endpoints;
@@ -149,8 +150,8 @@ public static class ReleasesEndpoints
 
         var releasesData = await reportRepository.GetReleases(from, to, releaseType, cancellationToken);
 
-        return CsvRequest.IsCsvRequired(httpContext)
-            ? CsvRequest.Result(releasesData.ToCsvResponse())
+        return Request.IsCsvRequired(httpContext)
+            ? Request.CsvResult(releasesData.ToCsvResponse())
             : Results.Ok(releasesData.ToResponse());
     }
 }
